fix(pipeline): use QCed FASTQ files for subsequent steps and save fastp.json in output subdir (#44, #45)
This commit ensures the pipeline uses the QCed FASTQ files produced by fastp in the alignment steps when starting from FASTQ input. It also fixes the issue where fastp.json was previously saved to the execution directory, now placing it correctly in the specified output subdirectory.

bumped version to 2.0.0-alpha.27.

Closes #44
Closes #45