feat: highlight coverage and integrate fastp metrics in final report
This commit implements threshold-based highlighting of mean VNTR coverage and integrates fastp output metrics into the final summary report with corresponding warnings and icons. Configuration for all thresholds now resides in the config file, and the generation of the summary report properly utilizes the provided config. It also fixes references to undefined variables and ensures the config dictionary is passed down correctly.

Bump version to 2.0.0-alpha.31

Closes #47
Closes #48